fix(接口测试): api定义的基本信息放到tab里&响应内容可以切换布局
